% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pcgr.R
\name{compile_xlsx}
\alias{compile_xlsx}
\title{Compile multiple sample data into one XLSX workbook}
\usage{
compile_xlsx(
  DATA_LIST,
  TITLE = "comp_PCGR_CPSR",
  OUTDIR = "./",
  DELIM = "\\\\."
)
}
\arguments{
\item{DATA_LIST}{list of data returned from find_parse_jsons}

\item{TITLE}{character string naming output XLSX}

\item{OUTDIR}{character string path to write file to (default: getwd())}

\item{DELIM}{character delimiter for parsing names from json files (default: '.')}
}
\value{
none, writes an XLSX workbook with worksheets per data type
        where the same GENOMIC_CHANGE is found, compile all samples containing into one line
}
\description{
Compile multiple sample data into one XLSX workbook
}
