% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pcgr.R
\name{find_parse_jsons}
\alias{find_parse_jsons}
\title{Find and parse jsons}
\usage{
find_parse_jsons(PARSE_DIR, DELIM = "\\\\.")
}
\arguments{
\item{PARSE_DIR}{character string path where JSONs are a held}

\item{DELIM}{character delimiter for parsing names from json files (default: '.')}
}
\value{
list of lists, each per sample for the data parsed in parse_jsons
}
\description{
Find and parse jsons
}
