% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pcgr.R
\name{parse_jsons}
\alias{parse_jsons}
\title{Parse PCGR, CPSR json.gz files for single sample}
\usage{
parse_jsons(PCGR_JSON = NULL, CPSR_JSON = NULL)
}
\arguments{
\item{PCGR_JSON}{pcgr file to be parsed}

\item{CPSR_JSON}{cpsr file to be parsed}
}
\value{
named list, based on VCF_SAMPLE_ID, of PCGR: snv, CNA, TMB, MSI; CPSR: SNV
}
\description{
Parse PCGR, CPSR json.gz files for single sample
}
