% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pcgr.R
\name{snv_reduce}
\alias{snv_reduce}
\title{Compile multiple sample data into one XLSX workbook}
\usage{
snv_reduce(DATA_LIST, WHICH)
}
\arguments{
\item{DATA_LIST}{list of data returned from find_parse_jsons}

\item{WHICH}{character string which list entry to work on: pcgr_snv, cpsr_snv}
}
\value{
tb of reduced SNVs
}
\description{
Compile multiple sample data into one XLSX workbook
}
